{{cookiecutter.service_name}}/requirements.txt
cached-property==1.5.1
certifi==2018.11.29
chardet==3.0.4
Django==2.2.4
django-cors-headers==2.4.0
django-leaflet==0.24.0
django-simple-history==2.7.0
djangorestframework==3.9.1
djangorestframework-gis==0.14
docker==3.7.0
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
jsonschema==2.6.0
psycopg2-binary==2.7.6.1
python-decouple==3.1
pytz==2018.9
PyYAML==4.2b1
requests==2.20.1
six==1.12.0
texttable==0.9.1
urllib3==1.24.2
websocket-client==0.54.0
djangorestframework-jwt==1.11.0
coreapi==2.3.3
Sphinx==1.8.4
